backend: Specifically fetch revision if lookup fails
Apparently sometimes GitHub has different merge heads in GitHub Actions
-- for the same PR! The git state the action sees and the git state you
see after clone are desynced somehow...

We now always fetch the revision directly if we could not find it, maybe
it is still around on the remote.

This has some unfortunate interaction with forks (we can now run *their*
commits too), but that should be fine for our usecase.