backend+executor: Return category for tests in results
This is not stored in the database (which it probably should), so we
synthesize it on the fly when returning tasks. Fine for now, but will
show no category for deleted tests. Might also be good though, as
category reassignments will be accurately reflected, but that could also
be confusing.