client: Kill whole process tree on timeout/error
This is a bit sad. Ideally we would unshare into a new pid namespace and
kill the cgroup, but that's work. We could also set a new session id,
but that is not supported in rust. Finally, we could set a new process
group, but that did not work for reasons I did not dig into.

We now fall back to good ol' n² enumerating all processes that are part
of the tree and live with race conditions. After we found the process
tree, we merrily kill everything in it. If the pid was reused in the
meantime we kill random bystanders, so let's just hope this does not
happen often.