frontend: Speed up test overview table by ~2s
Creating vue components seems to be a major bottleneck if you do it
often enough. Therefore, we now use the library against its will and
hard-code all columns in the template. Additionally, we do not use
TableRow, TableCell or Button, but inline their definitions. This speeds
up rendering by a few hundred milliseconds…